# Conflicts:
#	lib/di/injector.config.dart
#	pubspec.yaml

+++ lib/mash/data/remote/models/academic/class_details_model.dart
@@ -13,12 +13,12 @@ class ClassDetailsModel extends ClassDetailsEntity {
   factory ClassDetailsModel.fromJson(Map<String, dynamic> json) {
     return ClassDetailsModel(
       className: json['CLASS_NAME'],
-      classId: json['CLASS_ID'].toDouble(),
-      academicId: json['ACADEMIC_ID'].toDouble(),
-      noSeats: json['NO_SEATS'].toDouble(),
-      order: json['ORDR'].toDouble(),
-      mediumId: json['MEDIUM_ID'].toDouble(),
-      isActive: json['IS_ACTIVE'].toDouble(),
+      classId: json['CLASS_ID'],
+      academicId: json['ACADEMIC_ID'],
+      noSeats: json['NO_SEATS'],
+      order: json['ORDR'],
+      mediumId: json['MEDIUM_ID'],
+      isActive: json['IS_ACTIVE'],
     );
   }
 }
